added audio files
the "soundtrack" is a reference to the title screen of final fantasy VI.
the idea is this: as sisyphus climbs the mountain, the notes are individually faded in based on progress, culminating in the final scary chords as he falls back to the start, only to begin again.
i hope javascript can handle amplitude level modulations & looping audio well, as it was a lot more time consuming than anticipated editing these audio files at the sample level to ensure smooth loopability.