still not working properly
the rockDistance variable counts down to 0 after it exceeds 6000, but the end.wav SFX still does not play until AFTER rockDistance reaches 0, despite the code for it to play appearing before (line 205). additionally, the text for rockDistance does not update, even though i tell it to in the loop. curiously, it only displays properly in the console.
i added a setTimeout function to delay the running of the summit() code by 1 second, but it doesn't seem to be working either, for some reason. (line 219)