game framework in place.
my previous project, the Sisyphus clicker game, was based on the interplay between 3 values: 1 which stood for the progress of the player, 1 which stood for the amount of progress the player can make per action, and 1 which stood for the amount of progress lost over time/per action.
i am taking that basic idea and expanding the number of variables/interplay involved. there will also be a fail-state. i won't go into detail as to the game's intended design right now, but the idea is that it will be an "action clicker", where you must constantly keep track of multiple values while making quick choices.