added new functions & buttons
there will be an "objective" added later on. right now the mechanics and basic feel are being put in place. flow is this: spend money to upgrade your console and use your console to "hack". as you hack, you will begin getting traced. the progress of the trace, "traceLevel", is based on the "traceFactor", which goes up each time a hack (currently 'ACTION') is pressed.
next will be incorporating "bouncing the signal" and "installing proxies", which will affect how you are being traced.
if "traced" reaches 100%, the current game - in theory - is lost.
"memory" is another resource the player must purchase and use. it will be used for hacking as well as delaying the trace.
tuning of values will be handled later once the basic mechanics are in place, and then will incorporate the "objective" and various gameplay "twists" to make things interesting [theoretically].