working on html/css/layout/visuals; updated variables/math; clarified some code
will have to provide visual indicators for tracing that aren't numbers; too many numbers on screen, which by itself isn't a problem, but some numbers will be more important than others depending on the context. want to create a sense of urgency when the trace gets close, so will do something with the colour red.
other than that, display looks alright, and the game is functional.
the lack of clarity of urgency in the interface is actually an interesting aspect. it forces you to pay attention to what you are doing.
more visual feedback for the various actions will change the feel.
audio is another consideration which will come later.