formatting; number display
began playing with some of the variables to achieve a semblence of good feel.
will begin working on "ultimate objective".
i'm thinking your "HACK IT" button could be toggled to hack the banks (to amass funds), or to hack the CENTRAL MAINFRAME, the root off this world's undisclosed evil. or maybe you are the evil. anyways, that way, there would be a balance between going back and forth to accrue money to pay for things like more memory. also, when hacking the mainframe, getting traced becomes a bigger problem.
i will have to experiment with 'random events' - things that can happen that disrupt the player somehow.
ideas:
supply shortage: temporarily jacking the price of memory/processing upgrades.
rival systems raid: removes processing power temporarily.
improved trace: the enemy gets a fix on your location and the traceFactor is raised to a high degree for a short amount of time, forcing the player to buy proxies.
cyber-countermeasures: increases the amount of memory required to bounce the signal.
etc.