layout/display changes
will eventually have to change/modify/add to the proxy abilities, or perhaps add an option to upgrade the proxies currently installed.

might make tracing more aggressive at certain predetermined points in time, and that upgrading proxies adds a certain amount of "resistance" to it. so that bouncing connection would be more of a quick-fix whereas a proxy upgrade is more of a long-term investment.
will have to implement random events creatively in order to force players to make decisions between quick fixes and longer-term upgrades.