working on visual elements
before adding in events, i want to work on the visuals and tune the "feel". in order to facilitate a sense of urgency, timers will be added to the player's actions. the time it takes to complete an action will be represented by progress bars. therefore, if the player finds himself in a tough situation, they can't spam memory/proxies, encouraging players to play smart.

the table layout is a little bit less than ideal, but i am no pro in html/css so i will only end up changing it if i have time and everything else i want is implemented. maybe adding in the progress bars will help attenuate the layout issue. let's see!