included progress bar for signal bounce progress
also will be adding two new, bigger bars. one will determine the progress of you being traced, the other will keep track of the "main objective".
so basically, the game will be a "race" of sorts to see who fills the bar first.
hack the CENTRAL MAINFRAME before your location gets fully traced, or whatever. the downside to hacking the central mainframe over the banking establishments is that it does not earn you money, but adds to your trace factor. be careful!