proxy upgrade progress bar works; other changes
the traceFactor variable, which is the amount by which your trace changes each update(), reverts to to between 0 & 0.1 whether positive or negative. this means that if you do nothing with a positive trace factor, it will go down - and if you do nothing with a negative trace factor, it will go up. either way, you will end up with a very small positive trace factor (between 0 and 0.1)
this will probably be tuned when other mechanics are finished/put in place.