added HACK CENTRAL MAINFRAME button, aka the primary objective
game framework mostly in place. realized that once the central mainframe has been "hacked", a survival countdown (downloading files...) should be needed to win, and to provide a sort of climactic feeling. during this climax sequence, the trace factor will be raised, and if the player isn't prepared (by stocking up on the cash & memory necessary to fight back the trace, for example), they will lose during this final sequence.