game testing continues.
works, and feels OK. the trace bar turning red is a huge help, and it is balanced in such a way that if you get greedy with your hacking, you can give yourself another chance if you are quick with bounces & proxies.
lowered overall level of the music that plays halfway through the hack.