started assignment 8
i plan on doing a top down adventure game using Phaser, which i had initially disregarded but after looking into it more, i can see how it could be useful in creating something that aligns with my vision. i am currently in the process of learning the ins and out of using phaser as well as things like tilesets so i hope to have a framework complete by next week.