rewriting the createPlayer function
since i am using babel to compile, => is ES6 shorthand for an anonymous function and binding this to it. normally it would be something like:

function (obj) {}.bind(this);

this allows me to target the x & y values of the object in particular that i had set when i made my tilemap in Tiled. in this case, the starting position of the player, which is next to the bed.