player spawns as intended now!
re wrote the code to be less confusing in the future, changed the JSON files by adding more layers for return spawn points.
hopefully this will work with multiple entrances/exits per screen in the future, anticipate having to make more object layers in the tilemaps for it to function as intended.