YES! moving back through areas works as intended.
i wanted to code this efficiently, as in, not having to create a new scene for each map. however, with the current setup, i will have to check:
1. which map we are on
2. whether the exit leads to the next area or the previous one
every single time an exit is hit. since i plan on having about 70~ rooms (if they all remain 16x16 blocks, that is - i could make some maps bigger, if i wanted to, lowering the map count), i would have a 70-step if statement eg. map === 1, else if map === 2, else if map === 3 etc.
right now, i want to complete a rigid framework that doesn't crash, then i will move on to creating the rest of the world.
2 more foundational things need to be added, 1 i anticipate will be relatively simple (collecting items), and 1 i anticipate will be extremely complicated if i want it done exactly how i want (dialogue).