i have done it. i have displayed a map screen.
this was unbelievable frustrating. i have been working with a tutorial series, and was getting errors when losing my tilemap exported from Tiled using JSON. however, when i used the tilemap from the tutorial, it ran. what confused me further was that when i had tried opening the tutorial .json file and re-exporting it, i was getting the same error, despite the file being 100%, line-for-line, literally identical. after spending hours, HOURS troubleshooting this single issue, i decided to look at what version of Tiled was being used in the tutorial and hunted that down. used the old version to export the tilemap. now it runs. unbelievable.
i had to sacrifice A LOT of time on this singular issue.