returning works properly, however implemented in a dirty way
basically made another version of the first map (map1B) with a different player spawn, and referenced that when returning from map 2.
i know there's a better way than this lest i end up with hundreds of json tilemaps.
a better way would be to link specific doors to specific player spawns using Tiled custom parameters and then referencing them/lining them up in the json level files. i'll try and figure that out if i can, but if i can't in a timely matter, i'm going to move on because i'd really like to finish this game and there are other aspects i want to prioritize right now, mainly the dialog and item collection.