setting up NPC objects for dialog
i will have to learn how to implement plugins for displaying dialog in a tasteful way. there is a very specific thing i want to achieve with the messages in this game, and its effectiveness will come down to how the text is displayed.
this is probably going to be the most difficult thing to achieve in terms of reaching my vision for the game.