text box works, but it's very ugly.
also wasn't able to change the style of the text inside the text box, maybe a typo somewhere but it's still courier and i thought i'd removed all references to the courier font so i don't really understand why it's still like that.
very strange. i can hard code the text into the MessageBox prefab, but it won't accept style as an object instantiated in the world scene constructor.