decided to take a break from coding and did some tile/level design. made map 3.
 i'm deciding against my original idea of making the entire game out of individual 16x16 tilemaps. the areas will vary by size.

however, i have run into a strange issue in that map3's tileset doesn't load properly. i may have to remove all references to the tileset i was using before. it's just a pain whenever i want to edit a tileset, i have to re-do levels from scratch. and, if i programmed any tile animation in, it also has to be re-done. i will save it for a final step later on.