began to understand the camera a bit more
what was confusing to me was that i thought the game world size was a function of the camera, because depending on how you had your scene framed, you would not be able to walk outside the area that was initially visible upon entering a new map. it doesn't really matter now, but disabling world bounds has helped, and allowed me to have the camera zoomed in a bit more.
what confused me then was how the message boxes spawned: not by using the x y values of the canvas but of the actual game world.
since i have the camera following the player around, the dirty solution was to simply spawn the message boxes in relation to the player's position.