decided to re-do how i do maps.
the way i was doing the maps with the multiple instances of each map was really, really stupid of me. instead, i will be splitting the game up into multiple, much larger maps, so that i don't have to deal with multiple instances of each and bloated numbers of files and lists of referencing each map file.