cleaned up files, but running into prefab issues
i want to give all my sprites animations, but assigning animation based on a prefab seems to have placed that object's spritesheets over everything in the object layer, ie. the books spritesheet is imposed on both the books object and the sign1 object, instead of the sign1 object having its own spritesheet (which it does).
the dirty workaround is to create a new class for every item, which i might do, considering it also might make it easier to keep track of which item has been collected or not and whether or not to respawn it when returning to an area or not. or it won't. who knows.