re-designed act 1, and some thoughts on design.
should be working on finishing the game but i just love the design aspect.
the secret underground passage used to be more maze-like, with a literal maze in it, which honestly didn't make much sense - i just kind of put it there, just because. so i removed it and added some long winding dead ends that are meant to open up as a shortcut back.

the design of the secret underground passage is not meant to be complicated, reinforced by Max telling you that - if you think that it is - it's probably not.

what it is designed for is a kind of teasing of the artifacts. most if not all the artifacts you see in the secret underground passage are not obtainable at first, you must explore in order to find the path there. i use the term explore liberally here, as sometimes they're right behind the corner, and sometimes they're not. i don't want to make a LOOONG game, but i want to keep you intrigued from the get go, be it by exploration or item collection or simply unraveling the story.