working on final tree sequence
apologies in advance, the coding for the tree dialog is super sloppy. the way i am doing it right now was meant to test how it was presented in game, and then come up with a better way of programming it after i had tuned how it was presented.
it's ... kind of working. some dialogs don't show but there's probably typos in the code somewhere.