audio just doesn't want to behave.
why would it? i just want to stop an audio file playing. why can't it? because a new scene is created and the previously playing audio file becomes untouchable? so i have to stop it before creating a new scene? great! oh wait CANNOT READ PROPERTY OF STOP UNDEFINED, whatever that means. what a waste of time. this is so pointless.